+++ Sources/BraintreePayPal/BTPayPalClient.swift
@@ -48,8 +48,6 @@ import BraintreeDataCollector
     /// - Parameter apiClient: The API Client
     @objc(initWithAPIClient:)
     public init(apiClient: BTAPIClient) {
-        apiClient.shouldSendAPIRequestLatency = true
-
         self.apiClient = apiClient
         self.webAuthenticationSession = BTWebAuthenticationSession()