+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * BraintreeCore
   * Batch analytics events to FPTI
   * Re-use existing URLSession instance for `v1/configuration` and subsequent BT GW API calls
+  * Send `start_time`, `end_time`, and `endpoint` to FPTI for tracking API request latency
 
 ## 6.18.2 (2024-05-15)
 * BraintreePayPal